+++ proxy/tests/integration/common/mod.rs
@@ -6,7 +6,7 @@ pub mod proxy_builder;
 pub mod routing;
 
 pub use mock_control_plane::MockControlPlaneBuilder;
-pub use mock_runner::MockRunner;
+pub use mock_runner::{MockRunner, RunnerState};
 pub use proxy_builder::TestProxy;
 pub use routing::{
     insert_active_model, insert_active_model_multi, insert_active_model_with_metadata,